third_party/blink/web_tests/platform/generic/fast/scrolling/html-element-client-rect-excludes-scrollbars-expected.txt

Tests that the html element excludes scrollbars when reporting clientWidth and clientHeight, and that it implements scroll() by scrolling the frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS htmlElement.clientWidth is innerWidth - 20
PASS htmlElement.clientHeight is innerHeight - 20
PASS scrollX is 10
PASS scrollY is 10
PASS successfullyParsed is true

TEST COMPLETE
